Updated scriptmodule with major revision of the code
Made big change in scriptmodule to make it easier for ecwolf to load data files and for ease of adding mods. Removed the need for mercurial and replaced with downloadAndExtract from helpers file in RetroPie. Added keyboard escape fix patch .All of this was done with the help from cyperghost @crcerror.